Go Back Button extended to Object and Address Results
