+++ explorer/client/src/pages/address-result/AddressResult.tsx
@@ -5,6 +5,7 @@ import { useParams } from 'react-router-dom';
 
 import { ReactComponent as AddressIcon } from '../../assets/AddressIcon.svg';
 import ErrorResult from '../../components/error-result/ErrorResult';
+import GoBack from '../../components/goback/GoBack';
 import Longtext from '../../components/longtext/Longtext';
 import OwnedObjects from '../../components/ownedobjects/OwnedObjects';
 import TxForID from '../../components/transactions-for-id/TxForID';
@@ -30,30 +31,33 @@ function AddressResult() {
 
     if (addressID !== undefined) {
         return (
-            <div className={styles.results} id="textResults">
-                <div className={styles.addressid}>
-                    <span>
-                        <AddressIcon /> Address
-                    </span>
-                    <div id="addressID">
-                        <Longtext
-                            text={addressID}
-                            category="addresses"
-                            isLink={false}
-                        />
+            <>
+                <GoBack />
+                <div className={styles.results} id="textResults">
+                    <div className={styles.addressid}>
+                        <span>
+                            <AddressIcon /> Address
+                        </span>
+                        <div id="addressID">
+                            <Longtext
+                                text={addressID}
+                                category="addresses"
+                                isLink={false}
+                            />
+                        </div>
                     </div>
-                </div>
-                <div>
-                    <h1>Owned Objects</h1>
                     <div>
-                        {<OwnedObjects id={addressID} byAddress={true} />}
+                        <h1>Owned Objects</h1>
+                        <div>
+                            {<OwnedObjects id={addressID} byAddress={true} />}
+                        </div>
+                    </div>
+                    <div>
+                        <h1>Transactions</h1>
+                        <TxForID id={addressID} category="address" />
                     </div>
                 </div>
-                <div>
-                    <h1>Transactions</h1>
-                    <TxForID id={addressID} category="address" />
-                </div>
-            </div>
+            </>
         );
     } else {
         return <ErrorResult id={addressID} errorMsg={'Something went wrong'} />;
