+++ explorer/client/src/pages/transaction-result/TxResultHeader.tsx
@@ -1,15 +1,13 @@
 // Copyright (c) 2022, Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 import cl from 'classnames';
-import { useCallback } from 'react';
-import { useNavigate } from 'react-router-dom';
 
 import { ReactComponent as CallTypeIcon } from '../../assets/SVGIcons/Call.svg';
 import { ReactComponent as PublishTypeIcon } from '../../assets/SVGIcons/Publish.svg';
 import { ReactComponent as TransferObjectTypeIcon } from '../../assets/SVGIcons/TransferObject.svg';
-import { ReactComponent as ContentBackArrowDark } from '../../assets/SVGIcons/back-arrow-dark.svg';
 import { ReactComponent as ContentFailedStatus } from '../../assets/SVGIcons/failed.svg';
 import { ReactComponent as ContentSuccessStatus } from '../../assets/SVGIcons/success.svg';
+import GoBack from '../../components/goback/GoBack';
 import Longtext from '../../components/longtext/Longtext';
 
 import type { ExecutionStatusType, TransactionKindName } from '@mysten/sui.js';
@@ -44,17 +42,9 @@ function TxAddressHeader({ data }: { data: TxResultState }) {
     const statusName = data.status === 'success' ? 'success' : 'failed';
     const TxResultStatus = TxStatus[statusName];
 
-    const navigate = useNavigate();
-
-    const previousPage = useCallback(() => navigate(-1), [navigate]);
-
     return (
         <div className={styles.txheader}>
-            <div className={styles.txback}>
-                <button className={styles.longtext} onClick={previousPage}>
-                    <ContentBackArrowDark /> Go Back
-                </button>
-            </div>
+            <GoBack />
             <div className={styles.txtypes}>
                 <Icon /> {TxKindName}
             </div>
